fastapi-signup/fastapi-signup/pyproject.toml
[tool.poetry]
name = "fastapi-signup"
version = "0.1.0"
description = ""
authors = ["kkinter <ifol1129@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.110.0"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
sqlalchemy = "^2.0.29"
python-dotenv = "^1.0.1"
argon2-cffi = "^23.1.0"
pyjwt = "^2.8.0"
pydantic-settings = "^2.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
